Moving Shared Components into dedicated folder
