+++ src/app/app.module.ts
@@ -6,6 +6,9 @@ import { BrowserModule } from '@angular/platform-browser';
 import { RouterModule } from '@angular/router';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
+import { PageNotFoundPagesComponent } from './core/error/page-not-found-pages/page-not-found-pages.component';
+import { ProblemWithTheServicePagesComponent } from './core/error/problem-with-the-service-pages/problem-with-the-service-pages.component';
+import { ServiceUnavailablePagesComponent } from './core/error/service-unavailable-pages/service-unavailable-pages.component';
 import { FooterComponent } from './core/footer/footer.component';
 import { HeaderComponent } from './core/header/header.component';
 import { MessagesComponent } from './core/messages/messages.component';
@@ -87,20 +90,20 @@ import { SocialCareQualificationLevelComponent } from './features/workers/social
 import { SocialCareQualificationComponent } from './features/workers/social-care-qualification/social-care-qualification.component';
 import { WeeklyContractedHoursComponent } from './features/workers/weekly-contracted-hours/weekly-contracted-hours.component';
 import { WorkerSummaryComponent } from './features/workers/worker-summary/worker-summary.component';
+import { WorkersModule } from './features/workers/workers.module';
 import { YearArrivedUkComponent } from './features/workers/year-arrived-uk/year-arrived-uk.component';
-import { AutoSuggestComponent } from './shared/auto-suggest/auto-suggest.component';
-import { DatePickerComponent } from './shared/date-picker/date-picker.component';
-import { DetailsComponent } from './shared/details/details.component';
+import { AutoSuggestComponent } from './shared/components/auto-suggest/auto-suggest.component';
+import { DatePickerComponent } from './shared/components/date-picker/date-picker.component';
+import { DetailsComponent } from './shared/components/details/details.component';
+import { SubmitButtonComponent } from './shared/components/submit-button/submit-button.component';
+import { TermsConditionsComponent } from './shared/components/terms-conditions/terms-conditions.component';
 import { NumberDigitsMax } from './shared/number-digits-max.directive';
 import { NumberIntOnly } from './shared/number-int-only.directive';
 import { NumberMax } from './shared/number-max.directive';
 import { NumberPositiveOnly } from './shared/number-positive-only.directive';
 import { Number } from './shared/number.directive';
-import { SubmitButtonComponent } from './shared/submit-button/submit-button.component';
-import { TermsConditionsComponent } from './shared/terms-conditions/terms-conditions.component';
-import { PageNotFoundPagesComponent } from './core/error/page-not-found-pages/page-not-found-pages.component';
-import { ProblemWithTheServicePagesComponent } from './core/error/problem-with-the-service-pages/problem-with-the-service-pages.component';
-import { ServiceUnavailablePagesComponent } from './core/error/service-unavailable-pages/service-unavailable-pages.component';
+import { SharedModule } from './shared/shared.module';
+
 
 @NgModule({
   declarations: [
@@ -202,6 +205,8 @@ import { ServiceUnavailablePagesComponent } from './core/error/service-unavailab
         component: LoginComponent,
       },
     ]),
+    WorkersModule,
+    SharedModule,
   ],
   providers: [
     LocationService,
