+++ src/app/shared/components/details/details.component.scss
