+++ src/app/shared/components/terms-conditions/terms-conditions.component.html
